+++ src/services/routeServices/calculatedData/calculateTotalHeight.ts
