+++ src/services/routeServices/filters/filterByIncompleted.ts
@@ -1,4 +1,4 @@
-import type { Route } from "../../types/dataTypes"
+import type { Route } from "../../../types/dataTypes"
 
 
 export const filterByIncompleted = (routes: Route[]): Route[] => {
