+++ src/services/routeServices/filters/filterByRouteName.ts
@@ -1,4 +1,4 @@
-import type { Route } from "../../types/dataTypes";
+import type { Route } from "../../../types/dataTypes";
 
 export const filterByRouteName = (routes: Route[], routeName: string): Route[] => {
