+++ src/services/routeServices/orderBy/orderByMaxGrade.ts
@@ -1,4 +1,4 @@
-import type { Route } from "../../types/dataTypes";
+import type { Route } from "../../../types/dataTypes";
 
 export const orderByMaxGrade = (routes: Route[]): Route[] => {
     // Ordenar las rutas por routeGrade de mayor a menor
