+++ src/services/routeServices/orderBy/orderBySchool.ts
@@ -1,4 +1,4 @@
-import type { Route } from "../../types/dataTypes";
+import type { Route } from "../../../types/dataTypes";
 
 export const orderBySchool = (routes: Route[]): Route[] => {
     // Ordenar las rutas por routeSchoolIndex de mayor a menor
